Better handling of escapes in strings
The simple regex used to handle escaped characters in strings meant that
the escape character, \, was not directly matched to the escaped
character that followed, but was only recognized when preceding a double
quote character. In particular, parsing could break for strings ending
with an (escaped) slash.

Explicitly matching escape sequences forces the lexer to pair escape and
escaped characters, fixing the issue. The change doesn't otherwise
impose restrictions on what characters can be escaped.