Make utf-8 tests more portable
The tests used to fail in Windows as the character `å` in the .java files
was interpreted as `Ã¥`. `javac` was interpreting the UTF-8 encoded
files as WINDOWS-1252 encoded file.

By explicitly setting that UTF-8 encoding in the Gradle configurations
`compileJava` and `compileTestJava` we make sure that `javac` will
interpret all files as UTF-8 (which they already are).