Disallow identifiers that start with a number
First of all, this is a breaking change. JSLT code that uses identifiers
starting with a number will fail to parse. The only redeeming feature of
this change is that this is a common restriction in many programming
languages, which should somewhat limit the risk of mayhem.

The motivation for doing this change, and one which is reflected in the
tests, is how array slicing is picky about whitespace: `.data[1:2]`
would result in a cryptic error message, as the subscript was being
interpreted as an imported module's identifier:

  com.schibsted.spt.data.jslt.JsltException: Parse error: Encountered " "]" "] "" at line 1, column 10.
  Was expecting:
    "(" ...
     at <inline>:1:7
       (...stacktrace follows...)

With this change, all identifiers (and references to them) are required
to start with a letter or underscore. Dashes are similarly disallowed,
to avoid conflicts with the use of a minus operator now or in the
future.